Add integration tests and update existing test to use Testcontainers (#294)
* Add integration tests: one "heavy" one that uses a full Logstash
container and is only activated with a system property, and a
"lightweight" one that uses netcat to capture the log messages in a
lightweight container with netcat pre-installed. Both of these extend an
abstract base test, AbstractElkAppenderIntegrationTest.
* Update existing ElkAppenderFactoryTest to use the lightweight
container, so that the TCP tests don't throw exceptions when they get
"connection refused". This didn't cause the tests to fail, but is
confusing when you see a bunch of stack traces where you aren't
expecting them in successful tests.
* Add several test extensions to eliminate duplicated code across the
tests: LogstashContainerExtension and Dropwizard TestAppExtension.
* Add a logback-test.xml logging configuration for test code.
* Add several test dependencies including dropwizard-testing support,
dropwizard-core, kiwi-test, and testcontainers for junit jupiter.
* Remove the (now redundant) .keep file from src/test/resources

Closes #292
Closes #293