Add Javadocs to ElkAppenderFactory; add explicit no-args constructor (#312)
* Add class-level Javadocs to ElkAppenderFactory
* Add a no-args constructor and move all field initialization in there.
I did this mainly for readability; you can easily see all the default
values in the same place. Another reason is that you can easily
breakpoint inside the constructor during debugging.
* Add a test to verify the default values. To accomodate this, I added
package-private getter methods. The exception is includeCallerData,
because it overrides the superclass method.
* Don't allow setting the ElkLoggerConfigProvider. Had to add a Setter
annotation with AccessLevel.NONE since there is a class-level Setter
which generates setters for all fields.

Closes #310
Closes #311